+++ src/types.d.ts
@@ -1,0 +1,5 @@
+
+export type AnyValue = void | boolean | string | object | number | null;
+export type AsyncAnyValue = Promise<AnyValue>;
+export type AssertionFunction = () => AnyValue | AsyncAnyValue;
+export type AssertionValue = AssertionFunction | AnyValue | AsyncAnyValue;